Made long running create operations async
This will not block the UI thread and makes it appear that everything is
working faster. It has the consequence that you will see your files
appear in chunks (they will basically dribble in) instead of seeing all
your files checked out. It's a bit weird at first but I think it works
fine. You don't have to wait until all the files are checked out to
work. You can work with those which are available already.

@W-2294805@