Don't sync on "Deploy to Server"
Previously we used to sync with the attached org before deploying. However, because the sync was broken, it never did any syncing. Thus, you never got any sync messages.
Now that I've fixed the sync, a dialog pops up now. This differs from the old (unintended) behavior.
I'm going to follow the old (unintended) behavior because it has behaved that way for so long that it might as well be the intended behavior.

@W-2930227@