Handle the retrieve of Standard Objects
As part of the upgrade process to support Aura, we tighten the files
that we will refresh from a retrieve operation so that we only retrieve
the type that we sent. However, for StandardObjects, the actual type
that we receive back is actually CustomObject because you don't actually
specify its type as StandardObjects.

See
https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/manifest_samples.htm#manifest_standard_objects

So we need to actually account for this case and handle it.

@W-3437959@